+++ ntex/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [3.2.0] 2026-02-11
+
+* client: Add `middlewares` support for http client
+
+* web: Rename wrap() to middleware()
+
 ## [3.1.0] 2026-02-06
 
 * Fix apply_fn/apply_fn_factory usage